Consolidate UserController into PersonSearchController, some test coverage.
