+++ src/main/java/de/dm/activedirectory/client/rest/PersonSearchController.java
@@ -6,8 +6,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.security.Principal;
 import java.util.List;
 
 @Controller
@@ -19,8 +22,19 @@ public class PersonSearchController {
         this.activeDirectoryService = activeDirectoryService;
     }
 
-    @GetMapping("/person/search")
-    public ResponseEntity<List<Person>> serachForPersons(@RequestParam String searchQuery) {
-        return new ResponseEntity<>(this.activeDirectoryService.findPersonsByName(searchQuery), HttpStatus.OK);
+    @RequestMapping(value = "/person/search", method = RequestMethod.GET)
+    public List<Person> searchForPersons(@RequestParam String searchQuery) {
+        return this.activeDirectoryService.findPersonsByName(searchQuery);
+    }
+
+    @RequestMapping(value = "/user", method = RequestMethod.GET)
+    public Person getCurrentUser(Principal principal) {
+
+        List<Person> personList = activeDirectoryService.findPersonsByName(principal.getName());
+        if (personList.size() == 1) {
+            return personList.get(0);
+        }
+
+        return Person.builder().displayName(principal.getName()).build();
     }
 }
